requirements-dev.txt
fastapi
flasgger
SQLAlchemy
asyncpg
python-dotenv
uvicorn
greenlet
grpcio
grpcio-tools
git+https://github.com/capstone2025uoa/grpc.git
pytest
httpx